+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.0] - 2023-07-03
+
 ### Added
 - Method `fromArray` and `fromConfig` to graphql client.
 - A `SessionUser` to use for web auth instead of `GenericUser`.